Initial commit fo issue 2185 for OSX
